feat: Improve saving from and to temp folder
Saving models requires additional care, because of security concerns
with regard to temp folders. Secondly, the risk of using parameter files
not related to the model also has to be adressed. Finally, adding
documents can potentially decrease the accuracy of the search indices.
This is why a user prompt is added.