feat: Add cross-comparison for indexed files
This options helps answer the question: which of the queried documents
are most similar to each other? It will retrieve similarity scores only
for files included in your tree view selection. Compare against the
query option, which answers the question: which of the files included in
my corpus are most similar to my queried documents?