feat: Shuffle client dirs when iterating
With mid to large-sized corpora, iterating through dirs alphabetically
creates problems, because the dictionary is filled too quickly and is
less diverse. Shuffling client folders can improve this situation
by picking folders with different domains and main languages. Still, if
the problem persists, we might need to implement shuffling on a project
dir level as well.